+++ .travis/setup-pymysql.sh
@@ -1,0 +1,2 @@
+pip install -U pymysql
+`dirname $0`/mysql.sh