+++ .travis/setup-pypostgres.sh
@@ -1,0 +1,2 @@
+pip install -U pg8000
+`dirname $0`/postgres.sh